perf(acpx): defer Pi session catalog runtime
Punchcard-Session: coral-workshop-workshop-3f
