+++ extensions/acpx/src/pi-session-catalog-shared.ts
@@ -1,0 +1,3 @@
+export const PI_SESSIONS_LIST_COMMAND = "acpx.pi.sessions.list.v1";
+export const PI_SESSION_READ_COMMAND = "acpx.pi.sessions.read.v1";
+export const PI_TERMINAL_RESUME_COMMAND = "acpx.pi.terminal.resume.v1";
